requirements.txt
numpy==1.21.5
Pillow==9.1.0
timm==0.4.12
torch==1.11.0
torchvision==0.9.0
